+++ lib/referee/cli.rb
@@ -24,6 +24,7 @@ module Referee
                   prefix: nil,
                   output: default_output_directory,
                   error_on_missing_storyboard_ids: false,
+                  language: 'objc',
                   verbose: false }
 
       optparse = OptionParser.new do |opts|
@@ -41,6 +42,10 @@ module Referee
 
         opts.on('-e', '--error-on-missing-ids', 'Issue errors on missing storyboard IDs') do
           options[:error_on_missing_storyboard_ids] = true
+        end
+
+        opts.on('-l', '--language LANGUAGE', 'Toggle language between objc and swift') do |language|
+          options[:language] = language.downcase
         end
 
         opts.on('-v', '--verbose', 'Enable verbose mode') do
@@ -64,11 +69,14 @@ module Referee
     end
 
     def validate_config
+      is_swift = (@config.language == 'swift')
       output_valid = File.exist?(@config.output) && File.writable?(@config.output)
-      build_output.die 'Class prefix must be provided' unless @config.prefix
+      build_output.die 'Class prefix must be provided' unless (is_swift || @config.prefix)
+      build_output.info 'Class prefix will ignored...' if (is_swift && @config.prefix)
       build_output.die 'Output directory must exist and be writable' unless output_valid
       build_output.die 'An Xcode project is required!' unless @config.project
       build_output.die 'Xcode project must exist!' unless File.exist?(@config.project)
+      build_output.die 'Invalid language specified!' unless ['swift', 'objc'].index(@config.language)
       build_output.info 'Loaded and validated configuration...'
     end