+++ lib/referee/configuration.rb
@@ -1,7 +1,8 @@
 module Referee
   # Configuration container.
   class Configuration
-    attr_accessor :project, :prefix, :output, :error_on_missing_storyboard_ids, :verbose
+    attr_accessor :project, :prefix, :output, :error_on_missing_storyboard_ids, :verbose,
+    :language
 
     def initialize(config)
       config.each do |key, value|