+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and this file 
 ## [Unreleased][unreleased]
 ### Added
 - Ensure Gemspec has a minimum Ruby version. Choosing 2.x, as this is default on Yosemite machines.
+- Adds in Swift code generation via `--language swift`. Default is `objc`.
+  - Swift code generation will output a file `Resources.swift`.
 
 ## [0.1.1] - 2015-09-02
 Initial public release.